app/requirements/base.txt
Django==4.0.6
django-cors-headers==3.13.0
python-decouple==3.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
